move vfc hooks logic to dedicated folder `x/vfc`
