+++ x/vfc/hooks/hooks_test.go
@@ -1,0 +1,25 @@
+package hooks_test
+
+import (
+	"testing"
+
+	"github.com/dymensionxyz/dymension/v3/app/apptesting"
+	"github.com/stretchr/testify/suite"
+	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
+)
+
+type HooksTestSuite struct {
+	apptesting.KeeperTestHelper
+}
+
+func TestHooksTestSuite(t *testing.T) {
+	suite.Run(t, new(HooksTestSuite))
+}
+
+func (suite *HooksTestSuite) SetupTest() {
+	app := apptesting.Setup(suite.T(), false)
+	ctx := app.GetBaseApp().NewContext(false, tmproto.Header{})
+
+	suite.App = app
+	suite.Ctx = ctx
+}
